Fixed tail-call support, now lua scripts can make tail-calls without the c-stack/lua callstack growing. Added some optimizations of the Lua bytecode and generated machine code. Lowered the opcode limit from 800 down to 200 for the max function size that is JITed. Added support for specialized versions of each opcode function to support more optimizations.
git-svn-id: http://llvm-lua.googlecode.com/svn/trunk@39 f87b780f-1855-0410-a4d3-fd11cda3f415

+++ llvm-lua/compile.sh
@@ -3,7 +3,8 @@
 FILE=${1/.lua/}
 DEBUG="0"
 if [[ $DEBUG == "1" ]]; then
-	CFLAGS=" -O0 -ggdb "
+	#CFLAGS=" -O0 -ggdb "
+	CFLAGS=" -ggdb -O3 -foptimize-sibling-calls -fomit-frame-pointer -pipe -Wall "
 	LLC_FLAGS=" "
 else
 	#CFLAGS=" -ggdb -march=athlon64 -O3 -fomit-frame-pointer -pipe -Wall "
@@ -15,5 +16,7 @@ fi
 	llc $LLC_FLAGS --march=c -f -o ${FILE}_mod.c ${FILE}.bc && \
 	gcc $CFLAGS -o ${FILE} ${FILE}_mod.c liblua_main.a -lm -ldl
 
-rm -f ${FILE}.bc ${FILE}_mod.c
+if [[ $DEBUG == "0" ]]; then
+	rm -f ${FILE}.bc ${FILE}_mod.c
+fi
